Package dependencies better
Package libglfw and stb_image dynamic libraries for MacOS X (build on
Mavericks).  Also package DerelictStb_image which isn't available
elsewhere.

Update package.json to make it work on MacOS X as soon as you clone the
repo, and update the documentation to reflect the removal of config.sh
step.